refactor ( BJS2-1507(Get followers with filter) and BJS2-1513(Get followings with filter) ) {   some fixes }
