+++ src/main/java/school/faang/user_service/controller/subscription/SubscriptionController.java
@@ -3,8 +3,8 @@ package school.faang.user_service.controller.subscription;
 import jakarta.websocket.server.PathParam;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
-import school.faang.user_service.dto.filter.UserFilterDto;
 import school.faang.user_service.dto.user.UserDto;
+import school.faang.user_service.dto.user.UserFilterDto;
 import school.faang.user_service.service.subscription.SubscriptionService;
 
 import java.util.List;
