--- src/main/java/school/faang/user_service/dto/filter/FilterDto.java
@@ -1,4 +1,0 @@
-package school.faang.user_service.dto.filter;
-
-public abstract class FilterDto {
-}
