--- src/main/java/school/faang/user_service/filter/criteria/user/UserFilter.java
@@ -1,8 +1,0 @@
-package school.faang.user_service.filter.criteria.user;
-
-import school.faang.user_service.entity.User;
-import school.faang.user_service.filter.criteria.CriteriaFilter;
-
-public interface UserFilter<V> extends CriteriaFilter<User, V> {
-
-}
