+++ src/main/java/school/faang/user_service/filter/user/UserIdFilter.java
@@ -1,13 +1,13 @@
-package school.faang.user_service.filter.memory.user;
+package school.faang.user_service.filter.user;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.user.UserDto;
-import school.faang.user_service.dto.filter.UserFilterDto;
+import school.faang.user_service.dto.user.UserFilterDto;
 
 import java.util.stream.Stream;
 
 @Component
-public class IdUserInMemoryFilter implements UserInMemoryFilter {
+public class UserIdFilter implements UserFilter {
 
     @Override
     public boolean isApplicable(UserFilterDto userFilterDto) {
