+++ src/main/java/school/faang/user_service/service/filter/UserFilterService.java
@@ -3,22 +3,21 @@ package school.faang.user_service.service.filter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import school.faang.user_service.dto.user.UserDto;
-import school.faang.user_service.dto.filter.UserFilterDto;
-import school.faang.user_service.filter.memory.user.UserInMemoryFilter;
+import school.faang.user_service.dto.user.UserFilterDto;
+import school.faang.user_service.filter.user.UserFilter;
 
 import java.util.List;
 import java.util.stream.Stream;
 
 @Service
 @RequiredArgsConstructor
-public class UserInMemoryFilterService implements InMemoryFilterService<UserDto, UserFilterDto> {
+public class UserFilterService {
 
-    private final List<UserInMemoryFilter> userFilters;
+    private final List<UserFilter> userFilters;
 
-    @Override
     public Stream<UserDto> applyFilters(Stream<UserDto> userDtoStream, UserFilterDto filterDto) {
         if (filterDto != null) {
-            for (UserInMemoryFilter userFilter : userFilters) {
+            for (UserFilter userFilter : userFilters) {
                 if (userFilter.isApplicable(filterDto)) {
                     userDtoStream = userFilter.apply(userDtoStream, filterDto);
                 }
