+++ src/main/java/school/faang/user_service/service/subscription/SubscriptionService.java
@@ -1,7 +1,7 @@
 package school.faang.user_service.service.subscription;
 
 import school.faang.user_service.dto.user.UserDto;
-import school.faang.user_service.dto.filter.UserFilterDto;
+import school.faang.user_service.dto.user.UserFilterDto;
 
 import java.util.List;
