+++ src/test/java/school/faang/user_service/service/subscription/SubscriptionServiceImplTest.java
@@ -7,12 +7,12 @@ import org.mockito.InOrder;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import school.faang.user_service.dto.filter.UserFilterDto;
+import school.faang.user_service.dto.user.UserFilterDto;
 import school.faang.user_service.dto.user.UserDto;
 import school.faang.user_service.entity.User;
 import school.faang.user_service.exception.DataValidationException;
 import school.faang.user_service.repository.SubscriptionRepository;
-import school.faang.user_service.service.filter.UserInMemoryFilterService;
+import school.faang.user_service.service.filter.UserFilterService;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -35,7 +35,7 @@ class SubscriptionServiceImplTest {
     @Mock
     private SubscriptionRepository subscriptionRepository;
     @Mock
-    private UserInMemoryFilterService userFilter;
+    private UserFilterService userFilter;
 
     @InjectMocks
     private SubscriptionServiceImpl subscriptionService;
